+++ demos/world/src/main/java/is/codion/framework/demos/world/ui/CountryTablePanel.java
@@ -36,9 +36,8 @@ import java.awt.Dimension;
 final class CountryTablePanel extends EntityTablePanel {
 
 	CountryTablePanel(SwingEntityTableModel tableModel) {
-		super(tableModel, config ->
-						config.editable(attributes ->
-										attributes.remove(Country.CAPITAL_FK)));
+		super(tableModel, config -> config
+						.editable(attributes -> attributes.remove(Country.CAPITAL_FK)));
 	}
 
 	@Override