+++ changelog.md
@@ -35,6 +35,7 @@ Codion Change Log
 - EntityEditPanel, EntityTablePanel.setupControls() renamed configureControls().
 - MenuConfig renamed ControlConfig, create() renamed createControls(), now public.
 - EntityTablePanel.PopupMenuConfig and ToolBarConfig added, related refactoring.
+- EntityTablePanel, popup menu and toolbar configuration refactored, overriding no longer required.
 
 ## 0.17.42
 ### is.codion.common.core